Fix hello-world updates
