+++ packages/twenty-apps/hello-world/src/application.config.ts
